Begin to rework rfam code
Rfam is unlike all the other code in that it does not actually download
and store a file. It really should do that as it makes it much easier to
track down any problems after importing. Additionally having a single
'utils' file with all the code is silly. This begins to split the code
apart into logical units as and then will focus on parsing local files instead
of always making ftp connections.
