+++ luigi/databases/rfam/data.py
@@ -1,0 +1,225 @@
+# -*- coding: utf-8 -*-
+
+"""
+Copyright [2009-2017] EMBL-European Bioinformatics Institute
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+
+
+INFORMATIVE_NAMES = {
+    # The leading _ is to prevent this from hitting isrP
+    "_srp": "SRP_RNA",
+    "y_rna": "Y_RNA",
+    "hammerhead": "hammerhead_ribozyme",
+    "group-ii": "autocatalytically_spliced_intron",
+    "vault": "vault_RNA",
+    "tmrna": "tmRNA",
+    "rnase_p": "RNase_P_RNA",
+    "rnasep": "RNase_P_RNA",
+    "rnase_mrp": "RNase_MRP_RNA",
+    "telomerase": "telomerase_RNA",
+
+    # This is to prevent this from hitting 'ctRNA' as well
+    "^trna": "tRNA",
+
+    # These two patterns are so that we don't hit tracrRNA with rRNA
+    "_rRNA": "rRNA",
+    "PK-G12rRNA": "rRNA",
+    r"^RUF\d+$": "other",
+    '^CRISPR-': 'other',
+}
+
+SO_TERM_MAPPING = {
+    "SO:0000013": "scRNA",
+    "SO:0000077": "antisense_RNA",
+    "SO:0000252": "rRNA",
+    "SO:0000253": "tRNA",
+    "SO:0000274": "snRNA",
+    "SO:0000275": "snoRNA",
+    "SO:0000276": "precursor_RNA",
+    "SO:0000370": "other",
+    "SO:0000374": "ribozyme",
+    "SO:0000380": "hammerhead_ribozyme",
+    "SO:0000385": "RNase_MRP_RNA",
+    "SO:0000386": "RNase_P_RNA",
+    "SO:0000390": "telomerase_RNA",
+    "SO:0000404": "vault_RNA",
+    "SO:0000405": "Y_RNA",
+    "SO:0000454": "rasiRNA",
+    "SO:0000584": "tmRNA",
+    "SO:0000588": "autocatalytically_spliced_intron",
+    "SO:0000590": "SRP_RNA",
+    "SO:0000602": "guide_RNA",
+    "SO:0000644": "antisense_RNA",
+    "SO:0000646": "siRNA",
+    "SO:0000655": "other",
+    "SO:0001035": "piRNA",
+    "SO:0001244": "precursor_RNA",
+    "SO:0001463": "lncRNA",
+    "SO:0001641": "lncRNA",
+    "SO:0001877": "lncRNA",
+    "SO:0000593": "snoRNA",
+    "SO:0000392": "snRNA",
+    "SO:0000399": "snRNA",
+    "SO:0000594": "snoRNA",
+}
+
+RFAM_RNA_TYPE_MAPPING = {
+    "Cis-reg; leader": "other",
+    "Cis-reg; riboswitch": "other",
+    "Gene": "other",
+    "Gene; CRISPR": "other",
+    "Gene; antisense": "antisense_RNA",
+    "Gene; antitoxin": "other",
+    "Gene; lncRNA": "lncRNA",
+    "Gene; miRNA": "precursor_RNA",
+    "Gene; rRNA": "rRNA",
+    "Gene; ribozyme": "ribozyme",
+    "Gene; sRNA": "other",
+    "Gene; snRNA": "snRNA",
+    "Gene; snRNA; snoRNA; CD-box": "snoRNA",
+    "Gene; snRNA; snoRNA; HACA-box": "snoRNA",
+    "Gene; snRNA; snoRNA; scaRNA": "snoRNA",
+    "Gene; snRNA; splicing": "snRNA",
+    "Gene; tRNA": "tRNA",
+}
+
+DOMAIN_MAPPING = {
+    'RF00002': 'Eukaryota',
+    'RF00008': 'Eukaryota',
+    'RF00010': 'Bacteria',
+    'RF00011': 'Bacteria',
+    'RF00013': 'Bacteria',
+    'RF00014': 'Bacteria',
+    'RF00017': 'Eukaryota',
+    'RF00018': 'Bacteria',
+    'RF00021': 'Bacteria',
+    'RF00022': 'Bacteria',
+    'RF00023': 'Bacteria',
+    'RF00032': 'Eukaryota',
+    'RF00033': 'Bacteria',
+    'RF00034': 'Bacteria',
+    'RF00035': 'Bacteria',
+    'RF00037': 'Eukaryota',
+    'RF00038': 'Bacteria',
+    'RF00039': 'Bacteria',
+    'RF00040': 'Bacteria',
+    'RF00042': 'Bacteria',
+    'RF00043': 'Bacteria',
+    'RF00044': 'Viruses',
+    'RF00050': 'Bacteria',
+    'RF00066': 'Eukaryota',
+    'RF00077': 'Bacteria',
+    'RF00079': 'Bacteria',
+    'RF00080': 'Bacteria',
+    'RF00081': 'Bacteria',
+    'RF00082': 'Bacteria',
+    'RF00083': 'Bacteria',
+    'RF00084': 'Bacteria',
+    'RF00101': 'Bacteria',
+    'RF00106': 'Bacteria',
+    'RF00107': 'Bacteria',
+    'RF00110': 'Bacteria',
+    'RF00111': 'Bacteria',
+    'RF00112': 'Bacteria',
+    'RF00113': 'Bacteria',
+    'RF00114': 'Bacteria',
+    'RF00115': 'Bacteria',
+    'RF00116': 'Bacteria',
+    'RF00117': 'Bacteria',
+    'RF00118': 'Bacteria',
+    'RF00119': 'Bacteria',
+    'RF00121': 'Bacteria',
+    'RF00122': 'Bacteria',
+    'RF00124': 'Bacteria',
+    'RF00125': 'Bacteria',
+    'RF00126': 'Bacteria',
+    'RF00127': 'Bacteria',
+    'RF00128': 'Bacteria',
+    'RF00140': 'Bacteria',
+    'RF00162': 'Bacteria',
+    'RF00166': 'Bacteria',
+    'RF00167': 'Bacteria',
+    'RF00168': 'Bacteria',
+    'RF00169': 'Bacteria',
+    'RF00170': 'Bacteria',
+    'RF00174': 'Bacteria',
+    'RF00177': 'Bacteria',
+    'RF00195': 'Bacteria',
+    'RF00210': 'Viruses',
+    'RF00230': 'Bacteria',
+    'RF00234': 'Bacteria',
+    'RF00235': 'Bacteria',
+    'RF00236': 'Bacteria',
+    'RF00238': 'Bacteria',
+    'RF00240': 'Bacteria',
+    'RF00242': 'Bacteria',
+    'RF00243': 'Bacteria',
+    'RF00250': 'Viruses',
+    'RF00368': 'Bacteria',
+    'RF00369': 'Bacteria',
+    'RF00370': 'Bacteria',
+    'RF00371': 'Bacteria',
+    'RF00372': 'Bacteria',
+    'RF00378': 'Bacteria',
+    'RF00379': 'Bacteria',
+    'RF00380': 'Bacteria',
+    'RF00382': 'Bacteria',
+    'RF00383': 'Bacteria',
+    'RF00388': 'Bacteria',
+    'RF00391': 'Bacteria',
+    'RF00435': 'Bacteria',
+    'RF00436': 'Eukaryota',
+    'RF00442': 'Bacteria',
+    'RF00444': 'Bacteria',
+    'RF00456': 'Eukaryota',
+    'RF00480': 'Viruses',
+    'RF00489': 'Bacteria',
+    'RF00490': 'Bacteria',
+    'RF00503': 'Bacteria',
+    'RF00504': 'Bacteria',
+    'RF00505': 'Bacteria',
+    'RF00506': 'Bacteria',
+    'RF00512': 'Bacteria',
+    'RF00513': 'Bacteria',
+    'RF00514': 'Bacteria',
+    'RF00515': 'Bacteria',
+    'RF00516': 'Bacteria',
+    'RF00517': 'Bacteria',
+    'RF00518': 'Bacteria',
+    'RF00519': 'Bacteria',
+    'RF00520': 'Bacteria',
+    'RF00521': 'Bacteria',
+    'RF00522': 'Bacteria',
+    'RF00534': 'Bacteria',
+    'RF00552': 'Bacteria',
+    'RF00555': 'Bacteria',
+    'RF00556': 'Bacteria',
+    'RF00557': 'Bacteria',
+    'RF00558': 'Bacteria',
+    'RF00559': 'Bacteria',
+    'RF01502': 'Eukaryota',
+    'RF01570': 'Eukaryota',
+    'RF01849': 'Bacteria',
+    'RF01850': 'Bacteria',
+    'RF01851': 'Bacteria',
+    'RF01854': 'Bacteria',
+    'RF01855': 'Eukaryota',
+    'RF01856': 'Eukaryota',
+    'RF01857': 'Archaea',
+    'RF01959': 'Archaea',
+    'RF01960': 'Eukaryota',
+    'RF02540': 'Archaea',
+    'RF02541': 'Bacteria',
+    'RF02542': 'Eukaryota',
+    'RF02543': 'Eukaryota',
+    'RF02544': 'Eukaryota',
+}
